fastapi-news/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
requests-html
lxml_html_clean
python-dotenv
mysql-connector-python
groq
